Dockerized_Apps/cic-polyglot/requirements.txt
Flask
modbus-tk
pyopendnp3
pyserial
pyModbusTCP
